added check rule for method names suppressed security module as there are some classes that triggered the check and this module is soon to be decommissioned changed test method names moved the classes to gateway-common
