+++ api-catalog-services/src/main/java/com/ca/mfaas/apicatalog/ApiCatalogApplication.java
@@ -10,7 +10,7 @@
 package com.ca.mfaas.apicatalog;
 
 import com.ca.mfaas.enable.EnableApiDiscovery;
-import com.ca.mfaas.monitoring.LatencyUtilsConfigInitializer;
+import com.ca.mfaas.product.monitoring.LatencyUtilsConfigInitializer;
 import com.ca.mfaas.product.service.BuildInfo;
 import com.ca.mfaas.product.service.ServiceStartupEventHandler;
 import org.springframework.boot.SpringApplication;
