+++ codequality/checkstyle/suppressions.xml
@@ -3,4 +3,7 @@
     "-//Puppy Crawl//DTD Suppressions 1.1//EN"
     "http://www.puppycrawl.com/dtds/suppressions_1_1.dtd">
 <suppressions>
+    <suppress checks="MethodName"
+              files=".*security-module.*"
+    />
 </suppressions>
