+++ gateway-service/src/main/java/com/ca/mfaas/gateway/GatewayApplication.java
@@ -11,7 +11,7 @@ package com.ca.mfaas.gateway;
 
 import com.ca.mfaas.enable.EnableApiDiscovery;
 import com.ca.mfaas.gateway.ribbon.GatewayRibbonConfig;
-import com.ca.mfaas.monitoring.LatencyUtilsConfigInitializer;
+import com.ca.mfaas.product.monitoring.LatencyUtilsConfigInitializer;
 import com.ca.mfaas.product.service.BuildInfo;
 import com.ca.mfaas.product.service.ServiceStartupEventHandler;
 import org.springframework.boot.SpringApplication;
